APP/requirements.txt
tk==0.1.0
Pillow==9.2.0
googletrans==4.0.0rc1
pinyin==0.4.0
pandas==1.5.0
opencv-python==4.6.0.66
apache-beam==2.40.0
Cython==0.29.32
avro-python3==1.10.2
contextlib2==21.6.0
keras==2.9.0
lvis==0.5.3
matplotlib==3.6.0
pycocotools==2.0.4
pyparsing==2.4.7
scipy==1.9.1
tensorflow-io==0.26.0
tf-models-official==2.9.2
tf-slim
torch==1.12.1
torchvision==0.13.1
seaborn==0.12.0